Исрпавления по замечаниям ментора
