+++ __test__/gendiff.test.js
@@ -1,7 +1,7 @@
 import { test, expect } from '@jest/globals';
 import path from 'path';
 import fs from 'fs';
-import getDiff from '../src/genDifference.js';
+import genDiff from '../src/index.js';
 
 const getPath = (filename, format) => (
   path.join('.', '__fixtures__', `${filename}.${format}`)
@@ -14,6 +14,6 @@ describe.each(['stylish', 'plain', 'json'])('%s format', (presentFormat) => {
   test.each(['json', 'yml', 'ini'])('getDiffTest', (format) => {
     const pathToFile1 = getPath('before', format);
     const pathToFile2 = getPath('after', format);
-    expect(getDiff(pathToFile1, pathToFile2, presentFormat)).toEqual(result);
+    expect(genDiff(pathToFile1, pathToFile2, presentFormat)).toEqual(result);
   });
 });
