+++ index.js
@@ -1,0 +1,3 @@
+import genDiff from './src/index.js';
+
+export default genDiff;
