+++ src/buildTreeDiff.js
@@ -1,17 +1,4 @@
-import fs from 'fs';
-import path from 'path';
 import _ from 'lodash';
-import formatting from './formatters/index.js';
-import parse from './parsers/index.js';
-
-const getObjFromFile = (pathToFile) => {
-  const pathToFileNormalize = path.normalize(pathToFile);
-  const pathAbsolute = path.resolve(pathToFileNormalize);
-  const format = path.extname(pathAbsolute).slice(1);
-  const data = fs.readFileSync(pathAbsolute, 'utf-8');
-
-  return parse(data, format);
-};
 
 const getDiffProperty = (key, obj1, obj2) => {
   if (!_.has(obj1, key)) {
@@ -34,14 +21,8 @@ const buildTreeDiff = (obj1, obj2) => _.union(Object.keys(obj1), Object.keys(obj
     if (_.isObject(obj1[key]) && _.isObject(obj2[key])) {
       return { key, children: buildTreeDiff(obj1[key], obj2[key]), type: 'nested' };
     }
+
     return getDiffProperty(key, obj1, obj2);
   });
 
-const genDifference = (pathToFile1, pathToFile2, format) => {
-  const data1 = getObjFromFile(pathToFile1);
-  const data2 = getObjFromFile(pathToFile2);
-  const diff = buildTreeDiff(data1, data2);
-  return formatting(diff, format);
-};
-
-export default genDifference;
+export default buildTreeDiff;
