+++ src/helpers.js
@@ -4,6 +4,7 @@ const getFunction = (functions, type, errorText) => {
   if (!_.has(functions, type)) {
     throw new Error(`${errorText} '${type}'`);
   }
+
   return functions[type];
 };
